+++ src/app/error.tsx
@@ -1,4 +1,4 @@
 'use client';
 
 // eslint-disable-next-line no-restricted-exports
-export { default } from '@/components/Error/Error';
+export { default } from '@/components/Error';
