+++ src/app/page.tsx
@@ -1,3 +1,8 @@
+import { Fragment } from 'react/jsx-runtime';
+
+import AboutSection from '@/components/homepage/AboutSection';
+import EventsSection from '@/components/homepage/EventsSection';
+import HeroSection from '@/components/homepage/HeroSection';
 import constants from '@/constants';
 
 import type { Metadata } from 'next';
@@ -8,6 +13,12 @@ export const metadata = {
   title: `Homepage | ${constants.APP_NAME}`,
 } satisfies Metadata;
 
-const HomePage: Component = () => <h1 className="text-center">Coming Soon...</h1>;
+const HomePage: Component = () => (
+  <Fragment>
+    <HeroSection />
+    <AboutSection />
+    <EventsSection />
+  </Fragment>
+);
 
 export default HomePage;
