+++ src/backup/pages/NotFound.tsx
@@ -1,13 +1,12 @@
+import { usePathname } from 'next/navigation';
 import { useEffect } from 'react';
 
-import { useLocation } from 'react-router-dom';
-
 const NotFound = () => {
-  const location = useLocation();
+  const pathname = usePathname();
 
   useEffect(() => {
-    console.error('404 Error: User attempted to access non-existent route:', location.pathname);
-  }, [location.pathname]);
+    console.error('404 Error: User attempted to access non-existent route:', pathname);
+  }, [pathname]);
 
   return (
     <div className="min-h-screen flex items-center justify-center bg-gray-100">
