+++ src/data/homeAboutValues.ts
@@ -1,0 +1,34 @@
+import { Heart, Lightbulb, Target, Users } from 'lucide-react';
+
+const homeAboutValues = [
+  {
+    icon: Users,
+    title: 'Inclusive Community',
+    description:
+      'We welcome developers of all skill levels and backgrounds to learn and grow together.',
+    color: 'from-blue-500 to-blue-600',
+  },
+  {
+    icon: Heart,
+    title: 'Knowledge Sharing',
+    description:
+      'Our members actively share their expertise through talks, workshops, and mentorship.',
+    color: 'from-red-500 to-pink-500',
+  },
+  {
+    icon: Lightbulb,
+    title: 'Innovation Focus',
+    description:
+      'We encourage experimentation with cutting-edge technologies and creative problem solving.',
+    color: 'from-yellow-500 to-orange-500',
+  },
+  {
+    icon: Target,
+    title: 'Career Growth',
+    description:
+      'We provide networking opportunities and resources to advance your professional journey.',
+    color: 'from-green-500 to-emerald-500',
+  },
+];
+
+export default homeAboutValues;
