+++ src/data/pastEvents.ts
@@ -1,0 +1,28 @@
+const pastEvents = [
+  {
+    id: 4,
+    title: 'GraphQL with Node.js',
+    date: 'November 20, 2024',
+    attendees: 65,
+    rating: 4.8,
+    highlights: ['Live coding session', 'Q&A with experts', 'Networking'],
+  },
+  {
+    id: 5,
+    title: 'Microservices Architecture',
+    date: 'November 5, 2024',
+    attendees: 58,
+    rating: 4.9,
+    highlights: ['Case studies', 'Best practices', 'Hands-on workshop'],
+  },
+  {
+    id: 6,
+    title: 'Node.js Security Best Practices',
+    date: 'October 18, 2024',
+    attendees: 42,
+    rating: 4.7,
+    highlights: ['Security audit demo', 'Vulnerability testing', 'Real-world examples'],
+  },
+];
+
+export default pastEvents;
