+++ src/data/upcomingEvents.ts
@@ -1,0 +1,46 @@
+const upcomingEvents = [
+  {
+    id: 1,
+    title: 'Node.js Performance Optimization Workshop',
+    date: 'December 15, 2024',
+    time: '10:00 AM - 4:00 PM',
+    location: 'Tech Hub, Ahmedabad',
+    speaker: 'John Doe',
+    speakerRole: 'Senior Node.js Engineer',
+    description: 'Deep dive into Node.js performance optimization techniques and best practices.',
+    attendees: 45,
+    maxAttendees: 60,
+    tags: ['Workshop', 'Performance', 'Advanced'],
+    featured: true,
+  },
+  {
+    id: 2,
+    title: 'Building Scalable APIs with Express.js',
+    date: 'December 22, 2024',
+    time: '6:30 PM - 8:30 PM',
+    location: 'Co-working Space, SG Highway',
+    speaker: 'Jane Smith',
+    speakerRole: 'Full Stack Developer',
+    description: 'Learn how to build robust and scalable APIs using Express.js framework.',
+    attendees: 32,
+    maxAttendees: 40,
+    tags: ['Talk', 'API', 'Beginner'],
+    featured: false,
+  },
+  {
+    id: 3,
+    title: 'JavaScript Testing Strategies',
+    date: 'January 5, 2025',
+    time: '2:00 PM - 5:00 PM',
+    location: 'Innovation Center, Ahmedabad',
+    speaker: 'Mike Johnson',
+    speakerRole: 'QA Engineer',
+    description: 'Comprehensive guide to testing JavaScript applications with modern tools.',
+    attendees: 28,
+    maxAttendees: 50,
+    tags: ['Workshop', 'Testing', 'Intermediate'],
+    featured: false,
+  },
+];
+
+export default upcomingEvents;
